# docs/swagger/swagger.json
{
    "basePath": "/api/v1",
    "definitions": {
        "auth.RefreshTokenRequest": {
            "properties": {
                "refreshToken": {
                    "example": "deadc0de1234567890",
                    "type": "string"
                }
            },
            "type": "object"
        },
        "auth.SignInGuestRequest": {
            "properties": {
                "age": {
                    "example": 30,
                    "type": "integer"
                },
                "height": {
                    "example": 180,
                    "type": "number"
                },
                "weight": {
                    "example": 75.5,
                    "type": "number"
                }
            },
            "type": "object"
        },
        "auth.SignInRequest": {
            "properties": {
                "email": {
                    "example": "john@example.com",
                    "type": "string"
                },
                "password": {
                    "example": "SecurePassword123",
                    "type": "string"
                }
            },
            "type": "object"
        },
        "auth.SignUpRequest": {
            "properties": {
                "age": {
                    "example": 30,
                    "type": "integer"
                },
                "confirmPassword": {
                    "example": "SecurePassword123",
                    "type": "string"
                },
                "email": {
                    "example": "john@example.com",
                    "type": "string"
                },
                "height": {
                    "example": 180,
                    "type": "number"
                },
                "name": {
                    "example": "John Doe",
                    "type": "string"
                },
                "password": {
                    "example": "SecurePassword123",
                    "type": "string"
                },
                "weight": {
                    "example": 75.5,
                    "type": "number"
                }
            },
            "type": "object"
        },
        "response.Error": {
            "properties": {
                "errors": {},
                "message": {
                    "type": "string"
                }
            },
            "type": "object"
        },
        "response.Success": {
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
    },
...
    "info": {
        "contact": {},
        "description": "This is the API documentation for Swimo - a swimming management and tracking application.",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "termsOfService": "http://swagger.io/terms/",
        "title": "Swimo API",
        "version": "1.0"
    },
    "paths": {
...
        "/refresh-token": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Generate new access token using refresh token",
                "parameters": [
                    {
                        "description": "Refresh token request",
                        "in": "body",
                        "name": "request",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshTokenRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "examples": {
                            "application/json": {
                                "data": {
                                    "expiresIn": 900000,
                                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwidHlwZSI6InJlZnJlc2giLCJleHAiOjE1NjIzOTAyMn0.Vu6l8mHqZ8r6XxR2nV9f6Lk8pO6Lg7pO",
                                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiZXhwIjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                },
                                "message": "Token refreshed successfully"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired refresh token",
                        "examples": {
                            "application/json": {
                                "message": "Expired refresh token"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "examples": {
                            "application/json": {
                                "message": "Internal server error"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "summary": "Refresh JWT token",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/sign-in": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Authenticate user with email and password, returns JWT tokens",
                "parameters": [
                    {
                        "description": "Sign in request with user credentials",
                        "in": "body",
                        "name": "request",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignInRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Sign in successful",
                        "examples": {
                            "application/json": {
                                "data": {
                                    "age": 30,
                                    "email": "john@example.com",
                                    "height": 180,
                                    "name": "John Doe",
                                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                                    "weight": 75.5
                                },
                                "message": "Sign in successful"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "examples": {
                            "application/json": {
                                "message": "Invalid JSON Body"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "examples": {
                            "application/json": {
                                "message": "Invalid email or password"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "examples": {
                            "application/json": {
                                "errors": {
                                    "email": "Email is not a valid format"
                                },
                                "message": "Validation Error"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "423": {
                        "description": "Account locked",
                        "examples": {
                            "application/json": {
                                "message": "Account has been locked"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "examples": {
                            "application/json": {
                                "message": "Internal server error"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                },
                "summary": "Sign in user",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/sign-in-guest": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Authenticate guest user without credentials, returns limited access tokens",
                "parameters": [
                    {
                        "description": "Guest sign in request with optional user agent",
                        "in": "body",
                        "name": "request",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignInGuestRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Guest sign in successful",
                        "examples": {
                            "application/json": {
                                "data": {
                                    "age": 30,
                                    "expiresIn": 1799999,
                                    "height": 180,
                                    "name": "Guest",
                                    "refreshToken": "a065443670df3b9aec28d9c7e5e4b6b9de1d70a877954f26aef5c470437b81a3",
                                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJLaW5kIjoiZ3Vlc3QiLCJTZXNzaW9uSUQiOiJjMGJlODI1ZS02ZmZkLTRhNTgtOTVjYy00MDRkODQ2Njk5OGYiLCJTdWIiOiIiLCJJc3N1ZWRBdCI6MTc2MDU5ODU2NCwiRXhwaXJlc0F0IjoxNzYwNjAwMzY0fQ.KY0Q-Ws9B_-2QgPeNGwkLUrbrPwsR7_5kERUGnJkZqY",
                                    "weight": 75.5
                                },
                                "message": "Guest sign in successful"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "examples": {
                            "application/json": {
                                "message": "Invalid JSON Body"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "403": {
                        "description": "Guest sign in disabled",
                        "examples": {
                            "application/json": {
                                "message": "Guest sign in is currently disabled"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "429": {
                        "description": "Guest session limit reached",
                        "examples": {
                            "application/json": {
                                "message": "Guest session limit reached"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "examples": {
                            "application/json": {
                                "message": "Internal server error"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                },
                "summary": "Sign in guest",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/sign-out": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Revoke user session and invalidate JWT tokens",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully signed out",
                        "examples": {
                            "application/json": {
                                "message": "Sign-out successful"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "examples": {
                            "application/json": {
                                "message": "Unauthorized"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "examples": {
                            "application/json": {
                                "message": "Internal server error"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "summary": "Sign out user",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/sign-up": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Register a new user account with email, password, and profile information",
                "parameters": [
                    {
                        "description": "Sign up request with user details",
                        "in": "body",
                        "name": "request",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignUpRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "examples": {
                            "application/json": {
                                "message": "User registered successfully"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "examples": {
                            "application/json": {
                                "message": "Invalid JSON Body"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "409": {
                        "description": "Email already exists",
                        "examples": {
                            "application/json": {
                                "message": "Email already exists"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "examples": {
                            "application/json": {
                                "errors": {
                                    "email": "Email is not a valid format"
                                },
                                "message": "Validation Error"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "examples": {
                            "application/json": {
                                "message": "Internal server error"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                },
                "summary": "Sign up new user",
                "tags": [
                    "Auth"
                ]
            }
        },
...
    },
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}